update backEnd : materialized-path

+++ frontEnd/src/main/service/modelService.ts
@@ -9,38 +9,40 @@ import { baseService } from './base/baseService'
 import { Repository } from 'typeorm'
 import { repositoryUtil } from '../util/repositoryUtil'
 import { fileUtil } from '../util/fileUtil'
+import { existsSync } from 'fs'
 
 export class modelService extends baseService<modelNode> {
   private _modelNodeRepo: Repository<modelNode> = repositoryUtil.getRepository(modelNode)
   private _scriptDir = import.meta.env.MAIN_VITE_SCRIPTS_DIR
 
   // 注册模型
-  public register = async (data: modelNode, filePath?: string): Promise<Res> => {
+  public register = async (data: modelNode, parentId?: string, filePath?: string): Promise<Res> => {
     try {
       if (data.type === 'script' && !filePath) {
         return Res.error('script file is required for script type model')
       }
-
-      const model_node = new modelNode(data)
-
+      const new_model_node = new modelNode(data)
+      // 如果模型类型为script，则需要注册模型文件
       if (data.type === 'script' && filePath) {
         // 区分zip文件和脚本文件
         if (filePath.endsWith('.zip')) {
-          console.log('zip file')
           // 解压zip文件
-          await fileUtil.extractZipFile(model_node.id, filePath)
+          await fileUtil.extractZipFile(new_model_node.id, filePath)
         } else {
           // 复制脚本文件到模型目录
-          const modelDir = path.join(this._scriptDir, model_node.id)
+          const modelDir = path.join(this._scriptDir, new_model_node.id)
           await fs.mkdir(modelDir, { recursive: true })
-          await fs.copyFile(filePath, path.join(modelDir, model_node.model_config.entry_file ?? ''))
+          await fs.copyFile(filePath, path.join(modelDir, data.model_config.entry_file ?? ''))
         }
       }
-
-      const result = await this._modelNodeRepo.save(model_node)
-      return Res.success('model registered successfully', result.id)
+      // 注册成功后，保存模型节点
+      const result = await this.createNode(new_model_node, parentId)
+      if (result.success && result.data) {
+        return Res.success('model registered successfully', result.data.id)
+      } else {
+        return Res.error('Failed to register model')
+      }
     } catch (error) {
-      console.error(`Failed to register model: ${error}`)
       return Res.error('Failed to register model')
     }
   }
@@ -48,10 +50,14 @@ export class modelService extends baseService<modelNode> {
   // 删除模型
   public delete = async (id: string): Promise<Res> => {
     try {
-      await this._modelNodeRepo.delete({ id: id })
       // 删除模型目录
       const modelDir = path.join(this._scriptDir, id)
-      await fs.rm(modelDir, { recursive: true, force: true })
+      // 如果模型目录存在，则删除
+      if (existsSync(modelDir)) {
+        await fs.rm(modelDir, { recursive: true, force: true })
+      }
+      // 删除模型节点
+      await this._modelNodeRepo.delete({ id: id })
       return Res.success('model deleted successfully')
     } catch (error) {
       console.error(`Failed to delete model: ${error}`)
